pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "bleach>=6.2.0",
    "faker>=37.5.3",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-limiter>=3.12",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "markupsafe>=3.0.2",
    "psutil>=7.0.0",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "gevent>=25.5.1",
    "psycopg2-binary>=2.9.10",
]
